adding dedup for in cluster collectors
